% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pseudo_obs.R
\name{create_pseudo_obs}
\alias{create_pseudo_obs}
\title{Create pseudo observations}
\usage{
create_pseudo_obs(y)
}
\arguments{
\item{y}{numeric}
}
\value{
\eqn{Y_{ij}^*} as an vector with length \code{length(y)^2}.
}
\description{
The pseudo observations is defined as \eqn{Y_{ij}^* = I(Y_i \preceq Y_j)}.
}
\examples{
create_pseudo_obs(c(c(-1, 0, 1)))
}
