% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fun.R
\name{est_fun}
\alias{est_fun}
\title{PIM U-estimating function}
\usage{
est_fun(y, X, b, link, w = NULL)
}
\arguments{
\item{y}{numeric A vector of pseudo observations.}

\item{X}{matrix The pairwise difference of original design matrix.}

\item{b}{numeric The PIM parameter.}

\item{link}{character The link function: "logit", or "probit".}

\item{w}{numeric The weights, default is NULL.}
}
\value{
matrix A \verb{n^2 x p} matrix contains the \code{n^2} transposed estimating fun.
}
\description{
The component estimating function is
\eqn{\delta(b)=\frac{\dot{\mu}(x^T b)}{\mu(x^T b)[1-\mu(x^T b)]}[y-\mu(x^T b)]w x}.
Here, \eqn{\mu} is the inverse of the link function. In "logit" case,
it can be simplified.
}
