% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jac_est_fun.R
\name{jac_est_fun}
\alias{jac_est_fun}
\title{Jacobian of PIM U-estimating function}
\usage{
jac_est_fun(y, X, b, link, w = NULL)
}
\arguments{
\item{y}{numeric A vector of pseudo observations.}

\item{X}{matrix The pairwise difference of original design matrix.}

\item{b}{numeric The PIM parameter.}

\item{link}{character The link function: "logit", or "probit".}

\item{w}{numeric The weights, default is NULL.}
}
\value{
matrix A \code{p} by \code{p} Jacobian matrix of \code{res_fun(...)}.
}
\description{
The Jacobian has the generic form
\eqn{\dot{\delta}(b) = \frac{\partial}{\partial b^{T}} \delta(b) =
\frac{\left(2 \mu y-y-\mu^{2}\right) \dot{\mu}^{2}+\mu(1-\mu)(y-\mu) \ddot{\mu}}{\mu^{2}(1-\mu)^{2}} x x^{T}}.
}
