% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_design_matrix.R
\name{new_design_matrix}
\alias{new_design_matrix}
\title{New design matrix}
\usage{
new_design_matrix(X)
}
\arguments{
\item{X}{matrix original design matrix}
}
\value{
Assuming \code{n = NROW(X)}, \code{p = NCOL(X)},
the returned object is also an matrix but with dimension n^2 x p.
}
\description{
Create new design matrix \eqn{X_{ij}^* = X_j - X_i} for PIM.
}
\examples{
X <- matrix(c(4, 9, 2, 3, 5, 7, 8, 1, 6), ncol = 3, byrow = TRUE)
new_design_matrix(X)
}
