% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preceq.R
\name{preceq}
\alias{preceq}
\title{Preceq}
\usage{
preceq(x, y)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
\eqn{I(x \preceq y)}
}
\description{
Calculate \eqn{I(x \preceq y) = I(x < y) + 0.5 I(x = y)}
}
\examples{
preceq(x = c(0, 0, 0), y = c(-1, 0, 1))
preceq(x = 0, y = c(-1, 0, 1))
}
