% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich_meat.R
\name{sandwich_meat}
\alias{sandwich_meat}
\title{Meat of U sandwich estimator}
\usage{
sandwich_meat(y, X, b, link, w)
}
\arguments{
\item{y}{numeric A vector of pseudo observations.}

\item{X}{matrix The pairwise difference of original design matrix.}

\item{b}{numeric The PIM parameter.}

\item{link}{character The link function: "logit", or "probit".}

\item{w}{numeric The weights, default is NULL.}
}
\value{
matrix A \verb{p x p} matrix.
}
\description{
Meat of U sandwich estimator
}
