% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_pim.R
\name{summary_pim}
\alias{summary_pim}
\title{Get summary statistics for fitted PIM.}
\usage{
summary_pim(pim_obj)
}
\arguments{
\item{pim_obj}{\itemize{
\item fitted object from \code{\link[=pim_fit]{pim_fit()}}
}}
}
\value{
a data frame with columns
\tabular{ll}{
\code{term} \tab covariates \cr
\code{estimate} \tab point estimate \cr
\code{std.error} \tab standard error \cr
\code{statistic} \tab Z-statistic (asymptotic) \cr
\code{p.value} \tab p-value \cr
\code{link} \tab link function used \cr
\code{prob.index} \tab probabilistic index \cr
\code{prob.index.se} \tab standard error of the probabilistic index \cr
}
}
\description{
Get summary statistics for fitted PIM.
}
\examples{
pim_obj <- pim_fit(mtcars$mpg, mtcars[c("cyl", "disp")], link = "logit", nleqslv.global = "dbldog")
summary_pim(pim_obj)
pim_obj <- pim_fit(mtcars$mpg, mtcars[c("cyl", "disp")], link = "probit", nleqslv.global = "dbldog")
summary_pim(pim_obj)
}
