% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim_fit.R
\name{pim_fit}
\alias{pim_fit}
\title{Fit a probabilistic index model}
\usage{
pim_fit(
  y,
  X,
  link = "logit",
  w = NULL,
  init = NULL,
  tol = 1e-06,
  max.iter = 200,
  keep.data = FALSE
)
}
\arguments{
\item{y}{numeric The outcome vector.}

\item{X}{matrix The design matrix.}

\item{link}{character The link function: "logit", or "probit".}

\item{w}{numeric The weights, default is NULL.}

\item{init}{numeric The initial guess of Newton's method.}

\item{tol}{numeric The numeric tolerance of \code{nleqslv()}.}

\item{max.iter}{numeric The maximum iteration of Newton's method.}

\item{keep.data}{logical Should the returned object keep original data?}
}
\value{
A list containing the estimated coefficients and their covaraince
matrix. It also contains the diagnostics of \code{nleqlsv()} procedure.
If \code{keep.data} is \code{TRUE}, then the inputs \code{y}, \code{X}, \code{link}, \code{w}
will also be returned.
}
\description{
Fit a probabilistic index model
}
